CP-34467: Exchange certificate when a hosts joins a pool
After passing the pre-join checks the joiner and the pool exchange
certificates and place them in the filesystem, ready to be used. The
database is not modified for any of the two sides. When rebooting xapi's
startup code in the joining host will detect its internal certificate is
not in the database and will add it.

Current code does not try to merge the trusted certificates that exist
in the joining host and the pool.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>