CP-34470: Serve destination certificates when asked for a migration
When a cross-pool migration is being prepared, add to the metadata both
the pool master's (internal) certificate and the destination's host
certificate. This is done only when TLS certificates are needed, that
is, when TLS verification is enabled pool-wide.

They will be used by the sending host to connect to the pool and connect
to the xapi-nbd server in the destination host.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>