CP-39806: avoid opening Threadext modules
This allows us to sidestep most of the code in Threadext and use only
Mutex.execute, Delay, thread_iter_all_exns and thread_iter

This matters because Stdlib's Thread module is changing in ocaml 4.14
in advance of multicore, dropping the code from stdext will allow it to
compile in 4.14

Stdext's Mutex module only implements execute, the rest of functions
come from Stdlib's Mutex, so changing an open doesn't change the
behaviour.

Stdext's Thread has a whole different implementation from Stdlib, this
affect the xapi, message-switch, the idl scheduler and many components
rather pervasively, so this change should not be done lightly although
it doesn't look like semantically are different.

Since the uses of execute have to be changed in any case, I've opted to
use the name with_lock as I find it more self-explanatory than just
"execute".

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>