xapi-idl: Use mtime much more in the scheduler
Spans were already used in the scheduler, just not exposed in the
interface, which made some of their benefits limited. Now the
scheduler's users can use it as well and the scheduler code simplified

In particular the module for periodic yum metadata updates now is forced to
convert from Unix seconds to monotonic seconds. The conversion happened
previously but it was much harder to notice. The operation is not exact as the
Ptime manual warns, but since the update doesn't need to be exact to the second
it's fine as it is.

Scheduler tests have been changed to compare spans, allowing for errors
in measurements, instead of ad-hoc boolean tests

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>