CP-39894: (xenops-xc) explicitize pid locations for all daemons
These have been historically a key in xenstore, plus a pidfile, with
swtpm now it can be only a pidfile.

Now the shared code between the daemons takes this into account when
cleaning up pidfiles or needs to know whether the daemon is running.

All other code does not depend anymore on this and can use directly the
pidfile or the xenstore key if the daemon module latter publicizes
either.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>