database/redo_log: use timers for backoff delays and deadlines
Use a monotonic time counter to count time since the last time the
backoff delay was changed

Previously a failure timestamp of 0 meant there was no failure, but this
was never used, hopefully the code makes the basis of this code easier
to understand

Unixext has timed reads and writes using POSIX timestamps, so in this
case they can't be avoided and the timer provides a compatible interface as
well. It also makes it obvious that the conversion may be incorrect, this was
also the case previously, but it was more difficult that that was the case.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>