database/master_connection: use timers for retry timeouts
Now uses monotonic timers to ignore changes in the clock.

Previously a single variable was used to determine whether the retrying calls
have timeouts and which timeout they have, now they are independent.

Previously there was a default retry timeout and a normal retry timeout. The
first was ignored: it was as the first value of Master_connection.retry_timeout
but it was immediately overwritten to -1 and the default value was never used
again. The retry timeout behaviour is sometimes overriden by the system.

I haven't been able to find any evidence of uses of these conf values, and
because of the incorrect and complex behaviours I think these were only ever
used for development. The configuration for these has been removed.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>