CP-40357: Purge all removed fields from the database and clients
Now the database schema and most of the code generation are filtering
these. In the case where the documentation about removal is still
needed, the old method is used, albeit with a new name for the avoidance
of doubt.

This change means that it's impossible to create objects with removed
fields, introducing quite a bit of churn. The most salient issues here
are: the host metrics drop the free memory, which was still being
maintained and used for estimating evacuation time. This is now done
using total memory consumed by VMs instead; the removal
of metrics in VIFs and VBDs, which makes it impossible to delete these
objects from the database; and the removal of vmpp-related fields.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>