CA-379350: Use up-to-date vTPM UUIDs when creating device models
When creating a domain from an snapshot, the UUID of the vTPM as used by xapi
is changed as a new vTPM object is used. The xenctrl backend is not able to
fetch the new UUID as it only has access to the extra backend. Instead the UUID
is passed from the common xenops library, similarly to other devices.

Diagram of introduced behavior and in which cases these happen with analysis
below, it's assumed that xapi updates xenops metadata before doing any
lifecycle VM operation so the second column is always up-to-date with the
current xapi model.

   | extra vtpm | VM vtpm | result
   |------------|---------|-------
1. | None       | None    | None
   | -----------|---------|-------
2. | None       | Some b  | Some b
   | -----------|---------|-------
3. | Some a     | None    | Some a
   | -----------|---------|-------
4. | Some a     | Some b  | Some b
   | -----------|---------|-------
5. | Some a     | Some a  | Some a

1. Normal case when vtpm is unplugged, OK
2. This case happens in case of a xapi error, where it allowed hotplugging a
   vtpm device into a snapshot. Same with other resume cases, OKj
3. This case happens in case of a xapi error, where it allowed hot unplugging a
   vtpm in a snapshot. Same with other resume cases, OK
4. This is the expected result for snapshots. In other resume cases,
   hotplugging is disallowed by xapi, so this case shouldn't happen elsewhere.
5. Normal case when vtpm is plugged, OK

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>