uefi-certs: avoid breaking new VMs when override is set
Before this change, the use-case to enable custom certificates was:
1. Set override-uefi-certificates=true
2. Run xe-toosltack-restart
   - xapi deletes the symlink where varstored reads from and creates an empty
     directory
3. User calls pool.set_uefi_certificates

Between 2 and 3, any VMs created will have no UEFI certificates, breaking
secureboot for them

Furthermore, when a user wants to reset the remove custom certificates there is
no way to do so without restarting the toolstack and having set previously
override-uefi-certificates=false.

This path changes the behaviour (case 2b from Marcus' design) so the symlink is
maintained when the filesystem contents are synchronised from the database and
pool.uefi_certificates is empty. This ensure that between step 2 and 3, VMs get
a correct set of certificates; and allow users to clear the value of the field
and get the default certificates, without restarting the toolstack.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>
Tested-by: Samuel Verschelde <stormi-xcp@ylix.fr>