xapi-storage-script: Change directory creation
Avoid blowing up the stack when creating directories recursively.

Now an optimistic approach is used: create the directory, and if if cannot be
created try to create the parent. This avoid races in creation, but causes
using 2 calls per directory created for all the directories that need to be
created aside from the top-most.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>