xapi-stdext-std: change String.replace to replace characters
Now most of the implementation is shared with String.escaped

The few users that needed to replace strings, have been replaced with Astring's
cuts, as most of them were already segmenting strings, or they are run in very
specific, infrequent codepaths for efficiency to not matter.

Others have been replaced by Astring's filter as they were removing characters,
and the rest have been converted to the nre String.replace.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>