xapi-stdext-std: consolidate replace functions
Expose String.replaced and remove String.escaped, as it conveys better the
meaning of what the function does. Now Sexpr uses this new function for
escaping as well. The function String.escape has been left as is since it
provides a simple interface when only a single character needs replacing.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>