xapi-idl: Do not use custom String.explode.implode
These are highly inefficient.

Also changes some functions to be able to have less types and make normal usage
clearer. This comes at the cost of having to destructure the main type when
pattern-matching it.

Moves the device_number tests to its own executable to easily iterate on the
tests.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>