Clock: Use integers for date's timezone offsets
Previously the comments and code talked about "timezone printing", but that is
only a consequence of a more fundamental problem: datetimes without timezone
cannot be really compared to datetimes that have a timezone. This is because
the former do not have a common frame of reference with any other timestamp,
while the latter have the common frame of reference: UTC.

Now the code uses the exact offset in seconds, which also enables it to accept
datetimes that have an offset different than 0.

The invalid comparisons are not forbidden for the time being.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>