CP-40357: prepare datamodel for the removal of VMPP
The protection policy field of the VM can be removed.
The VMPP class is not easy to remove completely, it's really complex to
delete a class from xapi because of all the inbuilt assumptions.
The UUID field has been kept because the code generation assumes classes get
represented by ocaml records, but ocaml does not support empty records. At
least one must be present.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>