api_errors: add new error an use it: sr_feature_not_supported
Currently sr_operation_not_supported is used for two separate issues: when the
operation cannot be run at the moment, and when the SM backend won't be able to
run an operation.

Separate the two use cases by adding a new error for the second case, with the
feature always added to the error.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>