xapi-wait-socket: new service that waits for xapi to be available locally
This is needed in order to change xapi from finishing instantly to wait until
it has initialised completely. This is because we don't want the service that
can start running when xapi reaches this stage of initialization to wait until
it has fully initialised, which depends on storage being initialised, and hence
take a very long amount of time.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>