task_server: do not check for cancellation twice
When a cancellable task is run, the cancellation at the beginning needs to be
checked only once to stop the computation from happening and setting up the
callback so the computation can be notified of the cancellation.

It's unclear why the commit that added proper handling of cancellation before
the tasks starts did not remove the existing check.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>