task_server: do not check for cancellation twice
When a computation is task as part of task, the cancellation at the beginning
needs to checked only once to stop the computation from happening and setting
up the callback so the computation can be notified of the cancellation.

It's unclear why the commit that added proper handling of cancellation before
starting did not remove the existing check.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>