CP-53658, xenopsd: use domain_claim_pages on a single node, if possible
Xen currently supports to modes to claim memory for a domain: without any node
in particular, or claim memory in a single NUMA node.

When planning a domain, return the nodes that will host the domain, and how
much memory. In the case where the domain fits in a single NUMA node, claim
pages on that node, otherwise claim the memory without specifying any node.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>