network_utils: Allow adding a gateway that doesn't match any interface prefix
Currently when this happens, the ip command fails, but it's ignored, resulting
in nothing happening.

The onlink option allows to avoid the error and set the route anyway. This is
useful on cases where a cloud provider controls the gateway of the users'
hosts.

This argument was added unconditionally to debian's ifupdown in 2016 to avoid
this issue:
https://salsa.debian.org/debian/ifupdown/-/commit/8b7bca9597d2f75670b182f0fc149cdbaec3544c

This issue was reported by users, as well as in
https://xcp-ng.org/forum/topic/11160/off-subnet-off-net-gateway-configuration-routing-table-fails-with-xe-network-param-set

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>